requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
Django==4.1.7
django-appconf==1.0.6
django-celery-beat==2.4.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-easy-audit==1.3.5
django-extensions==3.2.3
django-ipware==6.0.2
django-rest==0.8.7
django-rest-framework==0.1.0
django-timezone-field==5.0
django-tracking-analyzer==1.1.1
django-user-agents==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
exceptiongroup==1.1.3
flake8==6.0.0
Flask==2.2.3
frozenlist==1.3.3
geoip2==4.8.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
kombu==5.3.1
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-text-splitters==0.0.1
langsmith==0.1.14
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.21.0
maxminddb==2.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
ngrok==0.2.2
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pathspec==0.11.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==3.0.1
PyJWT==2.6.0
PyPDF2==3.0.1
pypdfium2==4.27.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ipware==2.0.0
python-telegram-bot==20.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.3
stripe==5.2.0
telegram-send==0.34
tenacity==8.2.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2022.7
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.14
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.17.0
